requirements.txt
torch
setuptools
numpy
matplotlib
bitsandbytes
accelerate 
datasets 
transformers
tensorboard
peft
python-dotenv